Don't ask rummager for "topics"
`topics` can refer to two concepts in the GOV.UK stack:

1. Something in Whitehall we've renamed to "policy area"
(https://trello.com/c/KMOhcErO)
2. A kind of tag which used to be called "specialist sector"

The `topics` we are removing here refer to the whitehall-kind. We're
not using those in this application, so we can stop requesting these
from rummager. This will allow us to update rummager to rename `topics`
to `policy_areas` and `specialist_sectors` to `topics`.

Who said naming was hard?