Remove configuration of asset manifest file
This was used as a workaround for apps that have their assets compiled
once on Jenkins then rsynced to the servers. frontend has its assets
compiled on each machine, so this isn’t needed and we can safely let
Rails generate a manifest filename itself.

Tekin did an excellent writeup of the issue here:
https://github.com/alphagov/whitehall/commit/2f468927c5f8933cdbb7946786e
f664b0254383d