Use bold for search highlights
This is less intrusive than the yellow background highlighting previously
used, which is desirable because the highlights shouldn't get in the way
of reading the actual text.  Also, using a font weight instead of a
background colour means that the gaps between highlighted words aren't
visible.

The bold doesn't show up in titles, because titles are already bold.
However, keeping the markup in there seems desirable, because then we
can easily try out other markup options.  It also lets us build debug
tools to help us understand why things are matching the query.