Scroll to the selected tab on page load
This behaviour means that a selected tab retains the default fragment behaviour, i.e. people can copy links to a particular section of a site. An example url would be http://frontend.dev.gov.uk/view-driving-licence#other-ways-to-apply  - loading the page will jump to the tab strip with the correct tab selected and its content showing below.