Fix styling for link states with `mark`
Various combinations of link states were causing the text wrapped in `<mark>` to display incorrectly. This commit moves the link styles into the `h3` block as those are the only links that will have `<mark>` tags in. I have not set an `active` state as the `focus` state overrides this.