Remove code to modify search descriptions and titles
Frontend previously modified the titles and descriptions of search
results. To enable us to highlight the search description and title,
we've moved this logic to the publishing apps. For example, Whitehall
now correctly prepends "The home of HMRC on GOV.UK" to HMRC's search
description (https://github.com/alphagov/whitehall/pull/2314).

The truncation of the description is now the responsibility of rummager
(https://github.com/alphagov/rummager/pull/492).

Since turning on highlighting for all users on this app we no longer
need this special case code. However, in making this PR we've
discovered that for closed organisations, "Closed organisation:" is
prepended to the title. This will be fixed in a PR on Whitehall.

Trello: https://trello.com/c/GPInH4fj

Note that I've had to reformat the hashes in the test file because
Rubocop started complaining.