Remove special case for travel-advice before 2013-04-22
This special case was added when this feature was introduced to avoid
presenting unhelpful messages to the user while the content was being
updated.

Every published edition in the database has been updated after
this date, so there's no longer a need for this special case. Removing
this so as to simplify the upcoming migration work.