Log errors for Postcode states
When a user enters an invalid Postcode or a Postcode which has no match
in Mapit, we want to log this error. This commit modifies the code
specific to local transactions to log this error. It also changes some
of the code specific to licences that deals with the object being
returned from Mapit, as we now return a `GdsApi::HTTPErrorResponse` and
need to deal with this both in the action and `fetch_location` method.

This commit also includes some tests to make sure the instancr vars are
being set.