Add error logging for missing local interaction
We need to log the error when a local transaction has a local authority but no local interaction is available. The example is searching for postcode "DD5 3EW" on https://www.gov.uk/school-term-holiday-dates. The page already tells the user that no local interaction is available. This PR adds Rails logging and Google Analytics tracking for the new error code `laMatchNoLink`.

To log the exact message that the user sees when coming across this error, we refactor to move the text into the locale file and extend the `LocationError` class to support extra arguments that can be passed to the `I18n.translate` method.