Upgrade logstasher gem to add error notification
We log the postcode_errors with `logstasher` to JSON-style logs so that the errors can be viewed in Kibana. `logstasher` supports subscribing to ActiveSupport::Notifications, which we use to log the postcode errors.
https://github.com/shadabahmed/logstasher#listening-to-activesupportnotifications-events

We upgrade `logstasher` to release `0.6.2`. Initially we attempted to upgrade to the latest release `0.8.6`, but custom fields were not being output in the logs, notification subscription did not work, log messages for rendering view templates were being included and the JSON structure of the output was different (`fields` being deprecated since `0.6.5`). This appears not to be a straight-forward API change but rather changes in the internal architecture. In the interest of time I have settled on `0.6.2` as that includes the `watch` method and does not alter the previous logging output. I have logged an issue against the `logstasher` repo -> https://github.com/shadabahmed/logstasher/issues/92 to ask about the breaking change to notification subscriptions.

The commit adds the following to the JSON output:

`"postcode_error_notification":{
  "postcode_error":"invalidPostcodeFormat"
},`