Use stable version of Nokogiri
Nokogiri will be cutting an official release soon after more testing. This
hacked together release is causing some developers issues with building
Nokogiri, and the security vulnerabilities are no longer thought to be too
severe. Whilst we wait for an official release, we should revert as far as the
most recent stable release, 1.6.6.4, as this does address some security
vulnerabilites.