Change the "found the LA but have no service link" page
Using the prototype[1] as a guide we make this page a little more friendly
and less like an error condition.  We've also tweaked the controller and
views a bit to make use of the new local authority presenter which allows us
to not care which url the la has as long as there is one.