Don't use definite article in front of LA names
For most it doesn't sound correct (e.g. "the Sunderland City Council")
and for those where it does (e.g. "the Royal Borough of Greenwich") it
still sounds ok without and in one case the name already includes it ("The
Highland Council").

Where we said "We've matched this postcode to the <council>" we now say
"We've matched this postcode to <council>".  However, where we said "Search
the <council> website for this service" we leave the definite article in
because it's the website we're talking about.

This causes a complication because that text is part of a translation file
and we are limited in what logic we can perform.  We check if the la name
starts with a the in the LocationError and use a different translation that
excludes it for those councils whose names already include it.