Add GA tracking for results shown with data module
We implement GA tracking with data modules as described in govuk_frontend_toolkit (https://github.com/alphagov/govuk_frontend_toolkit/blob/master/docs/javascript.md#modules). We use the `auto-track-event` module defined in govuk_frontend_toolkit (https://github.com/alphagov/govuk_frontend_toolkit/blob/master/javascripts/govuk/modules/auto-track-event.js).

We also add the name of the first places results as a label for GA tracking. This allows for specific tracking of shown results in the DfE campaign to promote the "Report child abuse" page. For non-report-child-abuse cases this is not necessarily useful, but it is deemed as the best compromise to allow tracking for this specific campaign and may be extended in future (by sending all results) for other formats if useful.