Add functional tests for RootController to test CUSTOM hashes
We test that custom slugs and formats get recognised and passed into the render method. We wrap the RootController Constants in methods so that we can stub them in the tests, as per https://github.com/freerange/mocha/issues/13.

We wondered whether it would be better to test the RootController constants CUSTOM_SLUGS and CUSTOM_FORMATS directly with their preset keys and values (so without stubbing), but were unsure if it would make the tests harder to read.