Add  wide promo banner slot to the GOV.UK homepage
The homepage has a number of promotional slots, used to highlight
important things like dates (eg, self assessment), changes to policy
(eg, paper counterpart) or topical events (eg, budget).

These are quite smaller, and further down the the page, so don't work
as well for more prominent items. This adds a new full width promo slot
to try and improve visibility of high importance items.

It's not intended to be used permanently, but for the duration of important
events. Eg, things like elections, referendums.