Add publishing request id to travel advice index...
In order to trace content updates to travel advice comprehensively
we need to be able to track the publishing request and assert that
the content being rendered is related to the expected publishing action.
To do this render the publishing_request_id from the JSON data as a
meta tag on the travel advice index.