Allow multiple `filter_manual` fields in search
In service-manual, we need to pass two manual filters through to
rummager, to show the correct results. The code before only supports one
parameter being passed through, and will join more than one parameter,
for example if we have ["/service-manual", "service-manual"] we'll end
up with the string "/service-manual service-manual" in the hidden input.