Delete all references to local authority contact details
We used to use these as a final fallback option for local transactions if
there was no URL for the interaction or local authority, but stopped doing
that in 1dc18ad and have just stopped using the contact URL as a fallback
option too.

We're also removing the contact details from content API and Publisher now.