Stop falling back to contact URL after homepage URL
This isn't a useful fallback: if there's a contact URL for a local authority
then we should be able to use the root of that domain as its homepage. In
practice all local authorities in the LDG dataset [1] have both homepage and
contact URLs so this fallback is never used.

Removing this fallback means that we don't have to maintain an extra dataset
that isn't used here, and simplifies the code.

[1]: http://local.direct.gov.uk/Data/local_authority_contact_details.csv