Update AuthorityLookup to query mapit
As MapIt now provides slugs for local authorities, we can stop relying on the local authorities.json file. We remove the file and rewrite AuthorityLookup to query MapIt instead.

We also rewrite the AuthorityLookup class to make use of instance methods which makes the methods easier to divide into public and private methods (previously part of the class methods came after a private statements, but Ruby ignores this for class methods). We remove the gds code lookups as these are not used anywhere.