Add "b" content to the page as hidden div
This means the content for the "b" branch is rendered by the rails app rather than being in the JS implementation.  If it becomes complex we can use existing helpers and such to make sure everything is consistent.  It's hidden by default and only rendered for the register to vote page.