Remove duplication of lists of type codes
The previous commit identified that an area type had been included in one place
but not another. This helps a bit with that problem.

It's hard to check which ones are in the list if they're randomly ordered.

The ordering of the list of `authority_types` is important as licences need to
be able to select the most specific match in the case when two local
authorities are returned for a postcode. [This test describes the case](https://github.com/alphagov/frontend/blob/3b12a066f0787f9a926912becf04a8b440792f12/test/unit/licence_location_identifier_test.rb#L50L61)