Refactor LocationIdentifier subclasses
We refactor to make the intention of the LocationIdentifier subclasses more obvious. We also update to instantiate Objects so that we can make use of instance methods and variables, which cleans up the code. Previously the `areas`/`geostack` array was recast multiple times for different purposes which made it harder to follow the intention. We move most functionality into intention revealing methods.