Temporarily disable spelling suggestions
This disables the spelling suggestion feature.

In the last couple of weeks we've seen a couple of instances where
elasticsearch's spelling suggester returns an unfortunate replacement
for a valid search term (for example, "chillos" for "chilcot"). With a
ministerial reshuffle imminent, we'd like to avoid embarrassing
suggestions. When we know all the names of new cabinet members, we'll
add those to the ignore list in rummager, which will make sure that the
names won't get corrected.

The root cause of this issue is that elasticsearch suggestion feature
doesn't work well with the way we use it. I'll do a write up in
rummager about that.