Add search parameter to display the finder as open
The finder on the search results page is normally shown closed, unless at least one filter has been selected, in which case it is shown open. This adds a search parameter that forces the finder to be shown open even if no filters are selected. If the parameter is not present, then the logic on whether or not to display the finder reverts to the previous behaviour.

This feature has been added to allow searches originating from Whitehall (and possibly other applications in future) to always display the finder.