Remove references to superseded find your local council content item
As part of setting up the new 'Find your local council' page, we will remove this old content item which was a transaction. The new page will not be backed by a content item, so we remove all code referring to it. We update the tests to use a 'apply-blue-badge' fixture instead.