Add breadcrumbs and related links
The breadcrumbs and related items are set by slimmer (by `set_slimmer_artefact` in the ApplicationController) and takes a hash as parameter that looks like an artefact from content api. We can define a partial hash with the 'tags' key where we hardcode the breadcrumbs and related links. In future we may decide that we use a content item in the content-store to store this and other related data.

Note that to test the related links, we test for the slimmer test helper div `test-related`.