Use correct rails api for data attributes on register to vote promo
The rails API for defining data attributes on a tag doesn't walk through nested
hashes so if we want:

    <a href="http://hi.com" data-nested-attributes="please">Hi</a>

we might expect to be able to write:

    link_to 'Hi', 'http://hi.com/', data: { nested: { attributes: 'please' } }

but that doesn't work, and would instead produce:

    <a href="http://hi.com" data-nested="{attributes: \"please\"}">Hi</a>

Which is obviously wrong.  It will however translate underscores to dashes so we
are able to use ruby symbols as follows:

    link_to 'Hi', 'http://hi.com/', data: { nested_attributes: 'please' }

and this generates what we want.