Update GA tracking category for postcodeSearchStarted
We decide to update the track category for 'postcodeSearchStarted' from 'placeSearch' to 'postcodeSearch:<publication_format>'.
The word 'placeSearch' in the context of Frontend could be understood to refer to the 'find-my-nearest'/'place' format, but we track
'postcodeSearchStarted' for the following formats: place, local_transaction, licence and the newly created find_your_local_council.

We update the integration tests for all relevant formats.