Add error logging for location errors for find my local council
We aim to align the error handling for what's currently being done for LocalTransactions. This helps to further differentiate the types of errors the we expect from postcode search.

We add GA tracking by adding the data module to the postcode error summary tag. By doing this we also enable GA tracking for other formats as a side effect, but we believe this might be useful and can easily be filtered out. This also means we can remove the ga_tracking partial.

In order to be specific about the format which is triggering the GA tracking, we pass it as a separate parameter to the location form. There is already an attribute called `format` but this is being used to specify CSS class and does not match 100% with the names of the formats that we currently use.

Finally we reorganise the contexts in our integration test to make the different error cases easier to understand, which makes the git change quite noisy.