Add a survey link to the beta label for find local council
Each page (start, single result, two results) has a different survey so we use content_for to allow each page to pass the message back to the layout.

Because govuk_components are not rendered during tests we have to assert on the
json arguments to the component that are rendered instead.  This feels a bit
awkward.