Simplify homepage non js search box behaviour
Users without javascript will now see the label above the search box, rather than the label and input together next to each other within the white box.

This fixes a confusing situation for non javascript users where the entire search box looks like it should be clickable, but actually clicking the right side of the ‘box’ does nothing as the input field is not full width.