Add a .rubocop.yml file with todo's
This application is currently configured to only test changed files on
lint. When doing large scale refactoring (like #1004) this causes a lot
of noise, since you're forced to change a lot of unrelated code. It
makes the diff hard to read and is fairly frustrating, since linting
runs before tests when running rake.

Originally I intended to fix all the linting warnings in this app.
However, it's so creaky at some places that Rubocop's autocorrect will
actually break the tests.

Instead, this adds a .rubocop.yml file (generated with
--auto-gen-config). It allows us to fix some issues selectively, while
not introducing any more.