Generate breadcrumbs from content store
This replaces the slimmer generated breadcrumbs that used content api.

Some of the slimmer headers are no longer required since now it's only used to
render metadata and sidebars, and not every page has these.

We've also changed the behaviour of the travel advice index page so that it
returns 503 if a country content item is missing.  This is intentional, because
if countries dissapear we would like to know about it.