Set the related link for /find-local-council
The related links for /find-local-council are currently hardcoded
because the page doesn't exist in the content-api.

From https://github.com/alphagov/frontend/pull/1031 we will start using
the content-store to render the sidebar with related links. This means
we can set the links in the content-store and don't have to hard code
it.

Of course managing the links here is also a form of hardcoding. Once we
have an application that manages all related links (probably
content-tagger) we'll remove it here.