Remove special case for business support pages
Business support pages (like https://www.gov.uk/apprenticeshipsni)
explicitly don't have a sidebar. This is redundant, because the sidebar
won't show anyway when there's nothing to show (like related mainstream
pages and external links).

I haven't found any good reason in the commit history why this page is
a special case, so I think it would make sense to remove it. This means
that if we'd ever want something in the sidebar here we could make it
so.

🚫❄️