Don't raise error when content item is missing
Currently in most controllers we raise an error
(`GdsApi::HTTPNotFound`) when a page isn't found in the content store.

This was done so that we would know when we by mistake request the
wrong thing from the content store.

We made an exception for the rootcontroller, which is also used by
private-frontend, which displays unpublished content which may or may
not have a content item.

This commit now implements that behaviour everywhere, so that we're
more resilient to missing content items. In particular this fixes a bug
on private-frontend that prevents people from previewing simple smart
answers (https://govuk.zendesk.com/agent/tickets/1448901).