Add a new assisted digital satisfaction survey
This survey is specifically about help with fees and we only want to show
it on certain done pages.  For now this is a bit experimental, so we don't
want to add it to the documents themselves as a toggle selectable in
publisher.  We've left it as a hardcoded list in
lib/assisted_digital_satisfaction_survey.rb.

Note that the backend for this is handled by the feedback app.