Advanced Constraints for routing requests to format-specific controllers
To enable an incremental refactor of the RootController#publication
method, we want to split out responsibilities for individual formats
into their own controllers. This is done by an Advanced Constraint class
which makes a (cacheable) request to the content-api, and uses the
response to make a decision as to where the request will be routed.

In support of this new class is an Artefact Retriever Factory which
wraps the configuration needed to make a call to the content-api.