Update help format controller and test
We remove the deny_framing related methods as they only apply to the transaction and local_transaction formats.
We update the set_expiry method to only set the cache headers when live content is viewed.
We add tests to the controller tests for the recently created show page for both live and draft content. We also add an integration test for viewing content in draft.