Update answer pages to route to own controller
We use the AdvancedContraint to route all content of format 'answer' to their own AnswerController. We move over the tests from 'answer_rendering_test' to a more appropriately named file.
We had to update a routing test for SimpleSmartAnswers to stub requests to the ContentApi, as the index page now checks the format against the AdvancedConstraint defined for answer pages.