Add integration test for video format
We move over the small test from page_rendering_test and add tests to get better coverage for the video format, testing for preview and translation functionality. To test translation we had to add the updated_at field to the video fixture, so that it gets rendered.
We also removed the breadcrumb tests from answers and help format integration tests, as the line was duplicated in those tests.