Remove unused content security policy code
This was introduced in https://github.com/alphagov/frontend/pull/733.
It hasn't made it to production yet and we have no immediate plans to
do so. I really hope we pick this up again, but in the meantime I think
it's good to remove the code to avoid clutter and unnecessary
maintenance.

Allows us to remove one more thing from the deprecated
alphagov-deployment
(https://github.gds/gds/alphagov-deployment/pull/1289).