Remove custom set_expiry method for Publisher format controllers
The overridden `set_expiry` method would check if the request was a GET request. For all these controllers, the routes file only allows GET requests anyway, so that check was not necessary. This allows us to remove the method and just add the `viewing_draft_content` conditional check onto the before_filter.