Remove references to guide from RootController
The RootController tests commonly use the test helper `artefact_for_slug`. This helper sets the default format to be `guide`. Since removing the guide page from 'app/views/roots', a number of tests failed. We set the format for those failing tests to be either transaction or programme as those have not been extracted yet to their own controllers.