Clean up root controller and views
This removes root partials that are now present in the guide and programme views. We also remove the print functionality from the root controller as it is no longer required. Also we do not need to check if a part has been requested that does not exist, as only guides and programmes deal with 'parts'.