Update root controller test default format
By default, the mock artefact for the root controller test was for format 'answer'. We want to move away from testing formats that are not being handled by the root controller anymore (and the 'answer' format is now being handled by the AnswerController).
We replace the answer format with the business_support format, as that format has not been refactored out into a separate controller yet and will be the last 'simple' format to be refactored (besides the more complex and interactive 'licence', 'local_transaction' and 'place' formats).
Once the business_support format has been moved to its own controller, these tests will have to be revised in order to determine how useful they still are.