Update RootController to 404 for refactored formats
We have moved a number of formats to their own controllers. In certain edge cases, requests for these formats would not match their routes in the routes.rb file and end up in the RootController publication method (as this method deals all requests at the end of the routes files). We want to be more explicit and just return a 404 in case this happens.

The tests also add a test for a POST request for a refactored format. We've seen instances of inappropriate POST requests in Errbit for non-interactive formats such as the answer format. Since in our refactoring we no longer allow POST routes to get content (and we probably shouldn't) this used to return with a 500. Returning a 404 seems more appropriate. See [the comment in routes.rb](https://github.com/alphagov/frontend/blob/9037e30d09f7c72cc828a7c0615275712f97ed56/config/routes.rb#L69-L70)