Update travel advice controller and tests
As part of refactoring formats out of the RootController, we make some small changes to the travel advice format. We update the controller to redirect for api requests via a before filter, so that it matches the other controllers. We don't care about it setting cache-control headers for the redirect, as we don't do it for any of the other refactored format controllers.

We also update the integration test context descriptions to be more intention revealing. Previously there were two contexts with descriptions "a single country page preview" and "previewing a country page". It was difficult to understand the difference at first glance. Our confusion was in part due to the fact that Frontend is only used to render preview travel advice content. Actual content is already being rendered through multipage frontend.