Update RootController and tests
We can remove some code related to parts and associated tests from RootController as only Guide and Programmes deal with parts. The code still refers to params[:part] for LocalTransactions and Licences but we will remove that code once refactor those formats out in later PRs.

We also update the RootController test to remove any reference to the BusinessSupport format and replace it with appropriate alternatives. We remove a test about printing as that's another feature only related to Guides and Programmes, and we do not explicitly return a 404 for a print request, it just fails to route to any controller method. Therefore it feels inappropriate to test for that.