Remove post route for place format
When searching for a postcode in 'Find my nearest', a GET request is more appropriate since the request does not change any data. We have to update the location form to pass in the method as a parameter. We specify the method explicitly each time we create the LocationForm as we will aim to remove POST routes for the other unrefactored formats (LocalTransaction, Licence) as well.