Refactor PlaceController to remove CUSTOM_SLUGS
Previously we only used the CUSTOM_SLUGS hash for 'report-child-abuse'. It's easier to be explicit about this and we can remove various helper methods related to dealing with the CUSTOM_SLUGS hash. We initially tried to move the 'report-child-abuse' route to a new controller method but found that there would be a lot of code duplication in the Controller and Controller tests.

Instead we move all conditionals down into private method calls, making the controller method cleaner. In the process, we remove the postcode instance variable from PlaceController and rely on helper methods instead. This means we need to update the location form for the other formats.