Set postcode submit for place to be POST request
Previously we thought we could simplify the routing behaviour for postcode submission to be a GET request since we don't alter any data. However it's GOV.UK policy to not put confidential data such as postcodes into URLs as query parameters, so we need to revert to use a POST method. It's not ideal that the 'place#show' method now supports both GET and POST but it seems to be the most pragmatic solution at the moment. We also remove the method override on LocationForm as we will always want this to be a POST request to protect people's postcodes.