Update licence test to improve test coverage
We add more strict test expectations to describe what features we expect to see on the start page. This matches the tests of the other formats. We also add tests the bad postcode inputs repopulate the postcode field after submission and that previewing content is possible.

We also add the same location error handling as for LocalTransactions in the RootController to make the two formats more similar, and to make the error messages that the user sees more precise. Licences uses MapIt in the same way as LocalTransactions in that they use MapIt to convert postcodes into local authority slugs, so it makes sense to match the errors as well.