Update GdsApiAdapters to 38.0.0
There have been two major versions of the Api Adapters released since we
last upgraded this dependency.  The most signifiant change is that the
GdsApiAdapters::Response object no longer implements `method_missing`
as a delegator to an OpenStruct wrapper for the JSON response. In
practical terms, this means that we have to use the provided `:[]`
method for accessing the underlying JSON data, or explicitly ask for
the `parsed_content`.

One other relevent change is the deprecation of the Publisher API v1
adapter which involved the deletion of the `part_methods` helper. Here we
bring across the remainder of the deleted methods to the
`PublicationPresenter`.