Remove use of publication_path helper
We are removing the RootController. The PartHelper was still using the RootController's `publication` router helper to construct paths. This new implementation does not rely on a route helper (since we could use either guide_path or programme_path). We also update all other calls to the publication_path helper and create path helpers in routes.rb as required.