Refactor licence search method to make POST code path more obvious
Attempting to make it clearer at first glance what happens when a POST request is passed to the `search` method, as it's non standard to have a method support both GET and POST requests. We need to support POST requests for `search` so that we do not put a user's postcode into the URL.

We also update the radio button form for choosing from multiple authorities to a GET request since this information does not need be treated confidentially.