Catch invalid slug and return 404
Due to the unusual way in which this app parses URL fragments, it is
possible for it to construct an invalid URL for a backend resource.

Invalid URLs will be caught by the version of GdsApiAdapters referenced
here, and now the application controller can handle this case like any
of the other failure cases and return a cacheable 404.