Set publication instance variable as before_filter
This allows us to clean up more code. We also set the language as part of setting the publication instance variable. We didn't use to do this for every controller since for some we couldn't find any content items with Welsh language but it seems better to be consistent across all the Mainstream format controllers.