Update PlaceController to match other interactive formats
The place format is similar to LocalTransactions and Licences in that it is an interactive format, accepts a postcode from the location form an returns some location data. There is some benefit in make it more similar to the other two formats to reveal their similarities more clearly. The newly created ImminenceResponse encapsulates the error handling and is similar to the MapitPostcodeResponse class that we use for LocalTransactions and Licences. We also attempt to make the controller method clearer to read when a POST request is made.