Fix bug with scoped search results
In https://github.com/alphagov/frontend/pull/1082 the `SearchApi` class
was refactored. The internal `Searcher` class was inlined into
`SearchApi`. However, we didn't realise that `SearchApi` isn't a normal
class, but an object that is only instantiated in an initialiser.

This caused the instance variables to be persisted between requests,
which in turn caused very odd behaviour, since the scoped search
results would be remembered (but differently on different
machines/rails processes).

This makes improvements so that it can't happen again.