Fixes the NoMethodError: undefined method `places' error
The legacy of the Root Controller is a small handfull of controller
methods that handle both GET and POST HTTP requests. The PlaceController
is one of them, and we managed to introduce a small regression while
refactoring.

It seems that this page used to respond to postcode search
requests using GET, with the postcode in the query string. For reasons
of anonymity we prefer not to handle these requests, and would rather
that personal information is only submitted to the site using HTTP POST.

In our refactoring, we prevented this page from rendering correctly when
provided with a postcode in the query string. This change is to prevent
the page from erroring given same. As for any other unexpected query
string parameter, it will just ignore the provided info, rendering the
page as if the params were simply not there.