Add example page for A/B testing
Adds a new help page which acts as an example of an A/B test and
provides a way to manually and automatically check that A/B testing is
working. The main concern is that A/B testing cookies are being
correctly converted to custom headers by the CDN.

This simple example gives us a way to check A/B testing before we create
a real (and more complicated) A/B test of the navigation.