Add A/B test meta tag to A/B example page
This adds an easily-parsable tag to the page which specifies which A/B
bucket the user is in. The plan is to read this in the govuk Chrome
extension to make it easy for devs, testers and support staff to
identify which bucket they're in and switch between them.

Requests which go through the CDN will also pick up a cookie identifying
the A/B bucket, but on dev and integration (or when bypassing the CDN on
staging and production) we need an extra marker like this meta tag.